requirements.txt
Flask
selenium
pymongo
python-dotenv
requests
webdriver-manager